+++ frontend/webapp/containers/overview/index.tsx
@@ -2,3 +2,4 @@ export { OverviewContainer } from "./overview/overview";
 export { DestinationContainer } from "./destination/destination";
 export { SourcesContainer } from "./sources/sources";
 export { NewDestinationFlow } from "./destination/new.destination.flow";
+export { NewDestinationForm } from "./destination/new.destination.form";
